requirements.txt
asgiref==3.4.1
attrs==21.2.0
black==21.12b0
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
coverage==6.2
Django==4.0
djangorestframework==3.13.1
idna==3.3
iniconfig==1.1.1
model-bakery==1.3.3
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
pip==21.3.1
platformdirs==2.4.0
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.6
pytest==6.2.5
pytest-cov==3.0.0
pytest-django==4.5.2
pytz==2021.3
requests==2.26.0
setuptools==59.6.0
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.3
typing_extensions==4.0.1
urllib3==1.26.7
wheel==0.37.0
